requirements.txt
asgiref==3.8.1
Django==5.0.7
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
pillow==10.4.0
setuptools==65.5.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
django-modeltranslation